v2.5
Changed Structures and Added a function that user can upload a config file
