--- client/Dockerfile.bak
@@ -1,8 +1,0 @@
-FROM node:alpine
-
-WORKDIR /app
-COPY package.json .
-RUN npm install
-COPY . .
-
-CMD ["npm", "run", "dev"]
