+++ client/src/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
-// import App from './App'
 import reportWebVitals from './reportWebVitals'
 import LandingPage from './pages/LandingPage'
